dom/html/test/test_bug593689.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=593689
-->
<head>
  <title>Test for Bug 593689</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=593689">Mozilla Bug 593689</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 593689 **/
function testWidth(w) {
  var img = new Image(w);
  is(img.width, w|0, "Unexpected handling of '" + w + "' width");
}

testWidth(1);
testWidth(0);
testWidth("xxx");
testWidth(null);
testWidth(undefined);
testWidth({});
testWidth({ valueOf: function() { return 10; } });

function testHeight(h) {
  var img = new Image(100, h);
  is(img.height, h|0, "Unexpected handling of '" + h + "' height");
}

testHeight(1);
testHeight(0);
testHeight("xxx");
testHeight(null);
testHeight(undefined);
testHeight({});
testHeight({ valueOf: function() { return 10; } });

</script>
</pre>
</body>
</html>
